requirements.txt
streamlit
openai
llama-index
python-dotenv
PyPDF2
python-docx